chatbot prompt processing, skeleton backend, saved model
i think there is a problem with the saved model (mismatched keras versions in notebook vs. process python file maybe?)

problem with frontend- multiple generations cause the title of the page to jump down?

we need to add some sendData function from the dataStreaming file to send the data to the classifier in backend